Fixed Bug#4: Fire Station gets messed up and turns into Police Station
Whenever you built a fire station in the game, it would turn into a weird FD/PD hybrid tile. to fix it, i searched for where building the firestation was called. this led me to the MapScanner.java file where i found the doFireStation() function, where it was mistakenly calling the POLICESTATION instead of FIRESTATION; fixing this fixed the problem