Fixed Bug #5: Flooding can happen in the forest
The problem was that floods could happen anywhere, not just on the edge of a body of water. when i saw this problem, i realized that there must have been a problem in how the code defines the edge of a river. when i eventually found where this is defined (in TileConstants.java), I saw that there was an error in the isRiverEdge function - it classified the all of the types of WOODS & TREEBASE as being part of the rivers edge. when i fixed the isRiverEdge function to not include the woods and just the rivers edge, it works.