Fixed Bug #6: Building underwater wires costs the wrong amount
the problem was that the cost was the same for underwater wires and above ground wires. to find the error i knew that i had to find something that defined a wire cost and an underwater wire cost. when i searched for wire cost, it returned the roadliketool.java function that had to do with establishing the wire cost for the amount of wires placed. there was a case that checked if the wire was being placed underwater, and if it was satified it still returned the regular wire cost. once i changed that variable to underwater wire cost that was defined in the funciton, the problem was fixed.